Refactored views to a module
