+++ myapp/api/views/event_views.py
@@ -1,0 +1,1 @@
+# TODO: Move event-related view code here
