+++ myapp/api/views/user_views.py
@@ -1,0 +1,1 @@
+# TODO: Move user-related view code here
